hash get

+++ 0x1A-hash_tables/4-hash_table_get.c
@@ -9,7 +9,7 @@
 
 char *hash_table_get(const hash_table_t *ht, const char *key)
 {
-	unsigned int i = key_index((unsigned char *)key, ht->size);
+	unsigned long int i = key_index((unsigned char *)key, ht->size);
 	hash_node_t *node;
 
 	if (i >= ht->size)
